add link in courses model to user, and owner validation to courses controllers, fix typo in bootcamp.js (controller)

+++ controllers/courses.js
@@ -50,6 +50,7 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @access Private
 exports.addCourse = asyncHandler(async (req, res, next) => {
 	req.body.bootcamp = req.params.bootcampId;
+    req.body.user = req.user.id;
 	const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
 	if (!bootcamp) {
@@ -58,7 +59,15 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
 			404,
 		);
 	}
-
+    // Make sure user is course owner 
+    if(bootcamp.user.toString() !== req.user.id &&
+        req.user.role !== 'admin'
+    ) {
+        return next( 
+            new ErrorResponse(`User ${req.user.id} is not authroized to add a course to to bootcamp ${bootcamp._id}`,
+            401)) 
+    }
+    
 	const course = await Course.create(req.body);
 
 	res.status(200).json({
@@ -79,6 +88,14 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
 			404,
 		);
 	}
+    // Make sure user is course owner 
+    if(course.user.toString() !== req.user.id &&
+        req.user.role !== 'admin'
+    ) {
+        return next( 
+            new ErrorResponse(`User ${req.user.id} is not authroized to update course  ${course._id}`,
+                401)) 
+    }
 	course = await Course.findByIdAndUpdate(req.params.id, req.body, {
 		new: true,
 		runValidators: true,
@@ -102,6 +119,14 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
 			404,
 		);
 	}
+    // make sure user is course owner 
+    if(course.user.toString() !== req.user.id &&
+        req.user.role !== 'admin'
+    ) {
+        return next( 
+            new ErrorResponse(`user ${req.user.id} is not authroized to delte course  ${course._id}`,
+                401)) 
+    }
 	await course.deleteOne();
 
 	res.status(200).json({
